structifying, seperating enums, adding tests and test coverage library
