+++ Cargo.toml
@@ -9,3 +9,4 @@ edition = "2018"
 
 [dev-dependencies]
 rstest = "0.18.2"
+cargo-tarpaulin = "0.27.3"
