--- src/default/mod.rs
@@ -1,3 +1,0 @@
-mod contig;
-
-pub use contig::ContigDefault;
